% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemes-reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{theme_fivethirtyeight}
\alias{scale_color_fivethirtyeight}
\alias{scale_fill_fivethirtyeight}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggthemes}{\code{\link[ggthemes]{scale_color_fivethirtyeight}}, \code{\link[ggthemes]{scale_fill_fivethirtyeight}}, \code{\link[ggthemes]{theme_fivethirtyeight}}}
}}

